fix some bug

+++ src/net/sf/cherry/net/channel/handler/ItemSortHandler.java
@@ -1,9 +1,5 @@
 package net.sf.cherry.net.channel.handler;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-
 import net.sf.cherry.client.MapleClient;
 import net.sf.cherry.client.MapleInventory;
 import net.sf.cherry.client.MapleInventoryType;
@@ -13,10 +9,12 @@ import net.sf.cherry.server.MapleItemInformationProvider;
 import net.sf.cherry.tools.MaplePacketCreator;
 import net.sf.cherry.tools.data.input.SeekableLittleEndianAccessor;
 
-public class ItemSortHandler extends AbstractMaplePacketHandler
-{
-  public void handlePacket(SeekableLittleEndianAccessor slea, MapleClient c)
-  {
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+
+public class ItemSortHandler extends AbstractMaplePacketHandler {
+  public void handlePacket(SeekableLittleEndianAccessor slea, MapleClient c) {
     slea.skip(4);
     byte sort = slea.readByte();
 
@@ -29,14 +27,13 @@ public class ItemSortHandler extends AbstractMaplePacketHandler
     MapleInventoryType type = MapleInventoryType.getByType(sort);
     MapleInventory inventory = c.getPlayer().getInventory(type);
 
-    for (byte i = 0; i < 96; i = (byte)(i + 1)) {
+    for (byte i = 0; i < 96; i = (byte) (i + 1)) {
       if (inventory.getItem(i) == null) {
         continue;
       }
       if (inventory.getItem(i).getItemId() == 5110000) {
         c.getSession().write(MaplePacketCreator.enableActions());
-      }
-      else if (!items.contains(Integer.valueOf(inventory.getItem(i).getItemId()))) {
+      } else if (!items.contains(Integer.valueOf(inventory.getItem(i).getItemId()))) {
         items.add(Integer.valueOf(inventory.getItem(i).getItemId()));
       }
 
@@ -46,15 +43,15 @@ public class ItemSortHandler extends AbstractMaplePacketHandler
 
     MapleItemInformationProvider ii = MapleItemInformationProvider.getInstance();
 
-     for (Integer item : items) {
-     List<Byte> stack = new ArrayList<Byte>();
+    for (Integer item : items) {
+      List<Byte> stack = new ArrayList<Byte>();
       stack = inventory.findAllById(item.intValue());
       if ((stack.size() > 1) && (!ii.isRechargable(item.intValue()))) {
-        for (byte j = 1; j < stack.size(); j = (byte)(j + 1)) {
-          if ((inventory.getItem(((Byte)stack.get(j)).byteValue()) != null) && (inventory.getItem(((Byte)stack.get(j)).byteValue()).getQuantity() < ii.getSlotMax(c, item.intValue()))) {
-            for (byte k = 0; k < j; k = (byte)(k + 1)) {
-              if ((inventory.getItem(((Byte)stack.get(k)).byteValue()) != null) && (inventory.getItem(((Byte)stack.get(k)).byteValue()).getQuantity() < ii.getSlotMax(c, item.intValue()))) {
-                MapleInventoryManipulator.move(c, type, ((Byte)stack.get(j)).byteValue(), ((Byte)stack.get(k)).byteValue());
+        for (byte j = 1; j < stack.size(); j = (byte) (j + 1)) {
+          if ((inventory.getItem(stack.get(j).byteValue()) != null) && (inventory.getItem(stack.get(j).byteValue()).getQuantity() < ii.getSlotMax(c, item.intValue()))) {
+            for (byte k = 0; k < j; k = (byte) (k + 1)) {
+              if ((inventory.getItem(stack.get(k).byteValue()) != null) && (inventory.getItem(stack.get(k).byteValue()).getQuantity() < ii.getSlotMax(c, item.intValue()))) {
+                MapleInventoryManipulator.move(c, type, stack.get(j).byteValue(), stack.get(k).byteValue());
                 break;
               }
             }
@@ -65,14 +62,13 @@ public class ItemSortHandler extends AbstractMaplePacketHandler
 
     items.clear();
 
-    for (byte i = 0; i < 96; i = (byte)(i + 1)) {
+    for (byte i = 0; i < 96; i = (byte) (i + 1)) {
       if (inventory.getItem(i) == null) {
         continue;
       }
       if (inventory.getItem(i).getItemId() == 5110000) {
         c.getSession().write(MaplePacketCreator.enableActions());
-      }
-      else if (!items.contains(Integer.valueOf(inventory.getItem(i).getItemId()))) {
+      } else if (!items.contains(Integer.valueOf(inventory.getItem(i).getItemId()))) {
         items.add(Integer.valueOf(inventory.getItem(i).getItemId()));
       }
 
@@ -84,13 +80,13 @@ public class ItemSortHandler extends AbstractMaplePacketHandler
     for (Integer item : items) {
       List stack = new ArrayList();
       stack = inventory.findAllById(item.intValue());
-      for (byte j = 0; j < stack.size(); j = (byte)(j + 1)) {
+      for (byte j = 0; j < stack.size(); j = (byte) (j + 1)) {
         List new_stack = new ArrayList();
         new_stack = inventory.findAllById(item.intValue());
-        if (((Byte)new_stack.get(j)).byteValue() != current_slot) {
-          MapleInventoryManipulator.move(c, type, ((Byte)new_stack.get(j)).byteValue(), current_slot);
+        if (((Byte) new_stack.get(j)).byteValue() != current_slot) {
+          MapleInventoryManipulator.move(c, type, ((Byte) new_stack.get(j)).byteValue(), current_slot);
         }
-        current_slot = (byte)(current_slot + 1);
+        current_slot = (byte) (current_slot + 1);
       }
     }
   }
